Update patch set 20
Patch Set 20:

Right, I think a setting is the best way to handle it. I'm fine with the suggested config name, but it is verbose. We could also take the first sql connection if the config variable is unset (and optionally raise an error if there is more than one).

Patch-set: 20