Add angular to fetch-dependencies.sh
fetch-dependencies.sh will grab needed javascript libs. Now that the
zuul-web dashboard exists and uses angular, add angular to the script.

Version 1.5.8 was picked as that's what storyboard-webclient is using
currently, as well as what horizon/xstatic is using, so absent any other
inputs, it seems as likely to be the 'right' version as anything else.

Change-Id: I359a29784f210795a2c81d9e659dae798ff516f2