Update patch set 13
Patch Set 13:

Yeah, I think we're ultimately going to need to return all of the variants individually.  We're certainly going to need to do that when we start providing all of the detailed information about jobs.  So we may need an endpoint like /jobs/jobname/ which provides a list of variants for that job.  Meanwhile /jobs/ would really only provide the names, since the only thing all the variants have in common is the job name.

Having said that, we *could* cheat slightly and, as a convenience, provide the description for the first variant along with the job entry.  Generally speaking, If you had to pick a single description for a job, I think that's the one most likely to be accurate.

So /jobs/ returns [(name, description)]
and /jobs/jobname/ returns [(description, branches, files, ...)]

We can probably implement the second thing as a followup, as long as we know where we're going to put it in order to design the top level endpoints right.

Patch-set: 13