Update patch set 13
Patch Set 13: Workflow-1

Alright, that works for me, then maybe the web interface could have a drop-down table to list the variants and locations of them. I think that should be a follow-up because the job list is already useful as it is (e.g. user can discover and pick job from it when building their project's pipelines).

Do we need to change the route to support '/{tenant}/jobs/' and later '/{tenant}/jobs/{jobname}', or should we keep the '.json' extension we used in the other controller? (that would be '/{tenant}/jobs.json' and '/{tenant}/jobs/{jobname}.json'

Also, we should think about a nice job webpage where it could display the timeout, the number of nodes, the inputs/outputs, ... then that would translate into a '/{tenant}/jobs/{jobname}.json?details=true' which could be displayed when browsing to '/{tenant}/jobs/{jobname}.html'

-Workflow because this needs at least a break in 'job:list' to stop at the first description.

Patch-set: 13
Label: Workflow=-1