Update patch set 2
Patch Set 2:

I now understand the problem that this is meant to solve...

Before explaining, I'll clarify that there is a distinction in ansible between an *inventory* host and a *host* in the sense that you can have different *inventory* hosts that are really the same host underneath. Ansible doesn't care about that.

So you could have something like this in your inventory:

foo ansible_host=192.168.1.1
bar ansible_host=192.168.1.1
zuul ansible_host=192.168.1.1

These would be three distinct "hosts" from Ansible's perspective, even though in reality this is only one machine.

Now, the issue is when you have the same host in two groups, like so *AND* you have group_vars for each group *AND* you have conflicting variable names (the same variable declared in both group_vars):

    foo ansible_host=192.168.1.1
    
    [group_one]
    foo
    
    [group_one:vars]
    key=value

    [group_two]
    foo

    [group_two:vars]
    key=anothervalue

In this case, one of the two group_vars is going to necessarily win over the other. If nothing else is able to be used to resolve the conflict, the last fallback is based on the alphabetical order of the groups.

So, instead, we'd like do to this:

    foo ansible_host=192.168.1.1
    bar ansible_host=192.168.1.1    

    [group_one]
    foo
    
    [group_one:vars]
    key=value

    [group_two]
    bar

    [group_two:vars]
    key=anothervalue

So that there is no longer a conflict to resolve since the inventory hosts are different.

Patch-set: 2
Reviewer: Gerrit User 9061 <9061@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0