Update patch set 2
Patch Set 2:

What if, instead of mocking this out, we replaced lines like this:

http://git.openstack.org/cgit/openstack-infra/zuul/tree/tests/unit/test_scheduler.py?h=feature/zuulv3#n1588

with the framework you have here to actually run the CLI?

Currently the test I pointed to is validating the internal python API, which we don't care about.  The *actual* interface is the CLI which you're testing here, but these tests compare that to the internal gearman API, which is also subject to change.  A test that went from command-line to behavior would be ideal.

Patch-set: 2