Update patch set 2
Patch Set 2:

The motivation is to let operator change the tcp port (e.g. if for some reason the default one is blocked/reserved), and/or the file location of console file. It may be desirable to not pollute the /tmp of static host, and for the oci container, i'd like to be able to bind mount a shared location so that zuul_console can be served from the host instead of the slaves (e.g. /var/tmp/zuul-console/{log_uuid}.log).

Well, then it's a mitigation until zuul_console is removed.

Patch-set: 2
Reviewer: Gerrit User 9311 <9311@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0