Don't shrink windows on reconfiguration
The window size supplied by the user on a pipeline is really
the starting window size.  It can grow without bound, and so
there is no reason to shrink a window on reconfiguration.
For that matter, it's not necessary to increase it either,
unless the user has raised the floor higher than the current
value.

Therefore, use the existing window value on static (ie, dependent)
change queues during reconfiguration, or the new floor if it is
higher.

Dynamic queues will get new values from the pipeline (whose value
is also conditionally updated in the same manner).

Change-Id: Iaba25788ebe51ced919dc896aa20aa90675d7773