WIP: Git driver
This patch improve the existing git driver by adding
a refs watcher thread. This refs watcher looks at
refs added, deleted, updated.

When a refs is updated and that the related commit
include a change on .zuul.yaml/zuul.yaml or zuul.d/*.yaml
then tenants including that ref is reconfigured.

Furthermore the patch includes a triggering model
for the git driver. Events are sent to the scheduler
so jobs can be attached in the post pipeline for
the event ref-updated.

Change-Id: I529660cb20d011f36814abe64f837945dd3f1f33