Update patch set 2
Patch Set 2: Code-Review+2

This works for me. I have an OCD part of my brain which finds the need for multiple hosts with the same machine weird - but I think that's not my call to make. If we want Zuul to be able to create representations of the inventories needed for production, and some people might need this type of inventory for production, then I think this is a thing we should support - especially since we can support it easily.

Patch-set: 2
Reviewer: Gerrit User 2 <2@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2