Update patch set 2
Patch Set 2:

PS2 implements a solution to get the list of files changed for the HEAD of the ref-updated branch, via a new job call to the merger. The gerrit driver uses the Gerrit API to get the list of files changed for a change but there no way to do that with the git driver. So not sure that's the right way to do that but calling zuul/merger seems to work pretty well.

Patch-set: 2
Label: Verified=0