Update patch set 2
Patch Set 2: Code-Review-1

(7 comments)

In addition to the inline comments, I think we should start documenting the API.  We should probably put in a really big red note that this is a preview and is likely to change in backwards incompatible ways.  But now is the time to get starting on establishing how we document this so that we don't have to do it all at once later.

It would also be good to start getting some tests of the json endpoints.  They don't have to be fancy, but just exercising all of the code through the web server and the rpc server would be good.

Patch-set: 2
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1