Add implied-branches pragma directive
When two projects have dissimilar stable branch names, for example,
stable/pike and stable/jewel, but should generally be used together
and therefore share job variants, it can be difficult to make that
happen.  Currently, one must add explicit multi-branch matchers
to every such job and project-template.

This allows a user to add a pragma directive to indicate all the
jobs in a file should apply to multiple branches.

It also extends the pragma directive to cover the case of project
templates, which currently did not benefit from it, and need
explicit branch matchers even though the jobs adjacent to them
would not.

Change-Id: I57cf159992d8f501cbaf41aef19562951ef6b7ea