Rework log streaming to use python logging
python logging can log to a socket. If we open one as part of SSH
connection, we can get rid of the need for the zuul_console daemon. No
zuul_console daemon will remove a place where log streaming can fail,
and also will allow things to work properly after remote node reboots.

The log streaming is implemented as a thread in the executor so that we
can listen on an open port and then add forwarding that local port to
the remote node.

Change-Id: Ibf3e6d8a4c1c9f9d14c2741918d1795e7f920e55