WIP Rework log streaming to use logging
python logging can log to a socket. If we open one as part of SSH
connection, we can get rid of the need for the zuul_console daemon.

This isn't working at all, so it removes all of the tests from
.zuul.yaml to avoid wasting resources. I'm pushing it up in case anyone
else wants to run with it. The code added to
zuul/ansible/callback/zuul_stream.py is the code I was using as a local
test server to test using the log-to-socket mechanism from python.

Further work needed is:

* turn main() in to a thread that listens on the port
* add port a forwarding to the ssh connection

I think for the port forwarding, actually doing it in the executor
server and then adding something to the ssh options that get written out
would be best. That way the executor server can pick an open port, tell
ssh to forward that port to the remote hosts, and then tell the callback
plugin what local port to listen on.

It's also probably a good idea to do it in the executor server code
because the way we do this is almost certainly going to depend on the
connection driver.

It's also worth considering starting the logging listener thread as a
thread in the executor server and not in the callback plugin, and just
having it write to the same file. That said - since ULTIMATELY we'd like
to get streaming plumbed in to ansible core upstream, maybe it's worth
doing it in the playbook callback plugin for now.

Change-Id: Ibf3e6d8a4c1c9f9d14c2741918d1795e7f920e55