Add StandardJS linting and analysis
JavaScript Standard Style is a code style similar to pep8/flake8. It's
being added here not because of the pep8 part, but because the pyflakes
equivalent can catch real errors. This uses the babel-eslint parser
since we're using Babel to transpile already.

This auto-formats the existing code with:

  node_modules/.bin/standard --fix web/*.js

but otherwise does not make any functional changes. A follow up patch
will fix additional warnings/issues and switch from warn to error.

Change-Id: Ie55ee36aad3744c063ffe9e5fe61d86e13231229