Add StandardJS linting and analysis
JavaScript Standard Style is a code style similar to pep8/flake8. It's
being added here not because of the pep8 part, but becaue the pyflakes
equivilent can catch real errors. This uses the babel-eslint parser
since we're using Babel to transpile already.

This auto-formats the existing code with:

  node_modules/.bin/standard --global=jQuery --global=DemoStatusTree
    --global=DemoStatusOpenStack --global=DemoStatusBasic --global=URL
    --global=WebSocket --fix web/*.js

but otherwise does not make any functional changes. A follow up patch
will fix additional warnings/issues and switch from warn to error.

Change-Id: Ie55ee36aad3744c063ffe9e5fe61d86e13231229