Fix StandardJS warnings and turn them to errors
Most of these are underscore vs. camel case, which is boring and doesn't
matter at all. However, == vs. === is useful, as is ensuring template
strings are in backticks (which I actually got wrong in this series).

Fix both the stupid and useful stuff so we can get the useful stuff
caught for us.

Add a 'format' npm run target that will run the standard command with
--fix for ease of fixing reported issues.

Add a 'lint' npm run target and a 'lint' environment that runs with
linting turned to errors. The next patch makes the lint environment more
broadly useful.

Change-Id: Ia7b88ba11531a936699a7d883b57f9a6452bf7b0