Fix github app authentication in _getNeededByFromPR
When running as a github app we must supply the project in order to do
proper authentication as the installation. Otherwise we end up as an
unauthenticated request which will be rejected by most GHE
installations.

A draw back currently is that the search will be scoped by the app
installation (which is either a single repository or an
organization). As a future improvement we might want to iterate over
all installation and do the query. However this would need to be
configurable as with many installations this quickly can become quite
expensive. A different optimization could be searching in the local
change cache.

Change-Id: Iee7f009693dc8afa4631d2e48a630bb6da689aec