Make bundle of build web content
Building the files locally is wonderful and all, but what we really want
is to make a tarball that has the built code so that it can be deployed.

Put it in the root source dir so that it can be used with the zuul
fetch-javascript-tarball role.

Also, replace the custom npm job with the new build-javascript-content
job which naturally grabs the content we want.

Depends-On: I19607d14facb93d1f46ea68feba563badfbbb5a2
Depends-On: I9b96656aa5f225e4b069a78442cf0159cadc9009
Change-Id: I3108ea4151bc6a3d2fd87de6a163ee19397a5abb