Do not exit early tenant configuration in case of config issue
When zuul is starting all tenants are configured but if only
one job in one project is not well configured (example: reference
a unexisting job) then the whole tenant is broken.

When direct push to git repositories are allowed it is easy
for a user to push broken configurations then zuul
won't be able to restart successfully.

Change-Id: I935127d5a4ebb8eb0e3f9feeae59eb0e7f347c6e