Attempt to improve tenant configuration in case of config issue
When zuul is starting all tenants are configured but if
an issue happen like a project reference a non existing job
then the whole scheduler is broken.

When direct push to git repositories are allowed then
it is easy for a user to push broken configurations then zuul
won't be able to restart successfully.

This patch is not complete at all. Purpose is to check
if a suitable solution can be found.

Change-Id: I935127d5a4ebb8eb0e3f9feeae59eb0e7f347c6e