Make Zuul able to start with a broken config
When zuul is starting all tenants are configured but if
a bad config is detected in one tenant (like a project references
a non existing job) then the scheduler won't start successfully
and will remain in a broken state.

When direct push to git repositories is allowed then zuul
won't be able to prevent broken config to land. Users could push
broken configurations then zuul won't be able to restart
successfully.

This patch make zuul able to start with broken config and
register each detected config issues in an array attached
to tenants. This array of detected issue could be later
displayed by zuul-web to warn a tenant about issues related
to its zuul job config.

Change-Id: I935127d5a4ebb8eb0e3f9feeae59eb0e7f347c6e