Make Zuul able to start with a broken config
When zuul is starting all tenants are configured but if
a bad config is detected in one tenant (like a project references
a non existing job) then the scheduler won't start successfully
and will remain in a broken state.

When direct push to git repositories is allowed then zuul
won't be able to prevent broken config to land. Users could push
broken configurations then zuul won't be able to restart
successfully.

This patch makes zuul able to start with broken config and
register each detected config issues in an array attached
to the tenant.layout object. This array of detected issues
could be later displayed by zuul-web to warn a tenant about
issues related to its zuul job config.

The array of detected issue is limited to the first 10
issues detected by tenant.

Change-Id: I82c4595c2c73cf0192c15e00a8ffdff4a6cce0e6