Inject url endpoint information
Plumbing source_url and websocket_url around so that links work adds a
bunch of complexity to the codebase that's in service of dev/test. Add
a gate config that injects a json document as a global variable that we
can use to pull source_url and websocket_url information from.

As a followup, we can add a REST endpoint like /config.json that will return
these values, so the logic generally in the javascript can be "do I have a
global config var, cool, if not, GET config.json".

Change-Id: Ib63cc107f33a3fc5a0f2c5c2c3e15480d3d64419