Add run_ansible_setup_on_start flag for executor
On networking appliances, running 'setup' will just fail, as they
don't have Python.
I thought about maybe controlling that via the connection type plumbed
from Nodepool, but that would require creating a specific group in the
inventory for those nodes that setup will fail so we can later run pass
as pattern 'ansible all:!no-python' on runAnsibleSetup, but figured it
may be just easier to add it as a flag.

Change-Id: I112becbc3090ec8473c3a73ff1a2ff5f94658e09