Enable direct use of github driver in debug tool
The github debugging tool proved to be quite useful for prototyping
new stuff using the github api. In the process I ended up copying half
of the GithubConnection into it. However for trying out new stuff it
would be great to implement and try out the features/fixes directly in
the Github driver.

And here it is! The script now starts up a standalone version of the
Github driver which can be used to query and debug isolated stuff
against Github.

Change-Id: Ifdad1e69dae009011847869d51ff24000c44adb8