Add nodeenv and plumb it in to tox
For the case where a person doens't have JS tools installed, we can do
what we need to for the static web content completely in tox. However,
if someone does have JS tools installed, we don't want to muck with
things or take the cost.

For those who are adamant about not installing JS tools locally, now:

  tox -evenv -- npm run start

should work to do things with the js toolchain.

Change-Id: I330a947c7a824466ec24d089cd32728700ab4870