Add web server URL tests
This adds a framework for verifying that we can serve the web app
urls and their dependencies for all of the various ways we want to
support folks hosting zuul-web.

It includes a very simple reverse proxy server for approximating
what we do in openstack to "white label" the Zuul service -- that
is, hide the multitenancy aspect and present the single tenant
at the site root.

We can run similar tests without the proxy to ensure the default,
multi-tenant view works as well.

We should also expand this to test the sub-url scheme, that is,
mounting the service at a URL under the site root, such as what
software factory does.

Change-Id: Ib5c675bfddf9b234c0651ebf8576c055278c819e