Update patch set 2
Patch Set 2:

The nodaemon support basically just waits for receiving sigint and terminating then. This should also be handled here in ZuulDaemonApp (line 200). There the sigint and sigterm signal handlers are registered which take care of stopping the application.

With that the code paths of the app itself can be the same in daemon and nodaemon mode.

Patch-set: 2