Slack driver
This adds a slack reporter. It should also be the beginnings of a basis for
a slack trigger source.

The goal is to have a nice thread that encompasses all the runs for a
single queue item, rather than starting a new one each time. The cache
is in-memory, so it could grow rather large. However the
WeakKeyDictionary should cause it to remove cache entries as soon as
items are garbage collected.

Change-Id: I8a55be0f4c2e6a70fb5a98bada253dba2430e111