Update patch set 2
Patch Set 2:

That is an old patch I had on Wikimedia Zuul installation since July 2016. It caused a deadlock when running with  Paramiko before 1.11.6.

I don't think it is much of a problem nowadays since recent distributions come with a recent Paramiko version as well. Note how zuul/cmd/__init__.pp mentions to not import modules that import paramiko, so I guess we should merge this one as is.

Later we can get the Paramiko minimum version bumped from 1.8 to 1.11.6, move the imports back to the top of the files and drop the comments that refer to https://github.com/paramiko/paramiko/issues/59  But that is a little more work than this trivial patch.

Patch-set: 2