Update patch set 1
Patch Set 1:

I don't agree. We currently don't log anything for normal client connections and thus client disconnects should also be treated the same as they are part of normal client interaction.

If we want logging for each client connection (I'm not sure if we want this) this should be added in a separate change.

Patch-set: 1