Move CRD tests to test_gerrit_legacy_crd
This new class holds the unit tests which verify the soon-to-be
"legacy" form of cross-repo-dependencies which use the gerrit
change-id in the commit message footer.  The new form will use the
URL instead.  We will maintain both forms for some time, so keep
these tests around.  In a later change, this file will be duplicated
to exercise the same behavior with the new URL-based form.

Change-Id: I63a7398640bf82963fadf19acdcd2208c185da0b