Add job for testing against devel of ansible
By running Zuul's tests with tip of (or depends-on) of
ansible/ansible devel branch, we can know about any breaking changes
before they get released and either work with the ansible core team if
the break was unintentional, or work around it on our side.

List zuul in the required projects so that we can also run this same job
on ansible PRs in the future.

Change-Id: I05efffe37753e62778719602714a5f7e0291d2c4
Depends-On: I6d234247e3d45befd264e159c8b0fe77f01c7fdc