Use re2 for change_matcher
The regexes used in 'files' and 'irrelevant-files' are generated by
users. A bad set of backreferences and data can produce an extremely
expensive regex which will have so many permutations it will take longer
to resolve than the sun will take to supernova.

Since we want to keep letting users submit regular expressions, just not
with backrefs, we pivot to the much faster "re2" library for
change_matcher.

Other modules may be able to take advantage of re2's added speed, but
for now we'll target just the regular expressions which users might
accidentally get wrong.

This change may require some uses of backrefs to be moved to negative
match equivalents, such as changing files to irrelevant-files, or
branches to irrelevant-branches.

Change-Id: I5af229a86c2fe8f5541552b2a4dbeb716ee46d62