Update patch set 26
Abandoned

Hi,

Thank you for these changes to implement the GitHub driver.  We
did merge versions of many of these changes onto the
feature/zuulv3 branch.  That branch has now been merged back into
master in preparation for the 3.0 release.  Because that
functionality is now in the master branch, the versions of these
changes to master are no longer necessary so I am marking them as
abandoned (closed).

Thanks again for the contribution!

Patch-set: 26
Status: abandoned