Update patch set 13
Abandoned

Hi,

We recently merged the feature/zuulv3 branch back into master.
This includes many significant changes to the behavior and
underlying implementation in Zuul.  It appears this change has
either already been implemented in that branch, or is no longer
applicable.

Thank you very much for the contribution, and if this analysis is
in error, please accept my apologies and feel free to restore
this change and rebase it on the current tree.

Patch-set: 13
Status: abandoned