Tenant config can be read from an external script
This patch adds a change at tenant config file read level.
If the config file is detected as executable for the owner
it is then executed and the stdout is taken as the tenant config.
If the tenant file is not executable the standard yaml read is
performed.

The goal is to be able to fetch tenant configuration from
an external source whatever the type of the source as
the logic is handled in a script.

Change-Id: Ic8234dfcdf1f07f824542c4a2ebf99fe8ff19fd5