Tenant config can be read from an external script
This patch adds a change at tenant config file read level.
If a configuration option for the scheduler, tenant_config_script,
is set then the script is executed and the stdout is taken as the
tenant config.

The goal is to be able to fetch tenant configuration from
an external source whatever the type of the source as
the logic is handled in a script.

Change-Id: Ic8234dfcdf1f07f824542c4a2ebf99fe8ff19fd5