Tenant config dynamic loading of project from external script
This is a WIP. This patch adds a new format
of project definition that is a dictionnary with two
attributes path and function. When detected the config
loader will run the external script function to
generate the list of project.

The goal is to be able to fetch project list from
an external source whatever the type of the source as
the logic is handled in the python script.

Change-Id: Ic8234dfcdf1f07f824542c4a2ebf99fe8ff19fd5