Fix AttributeError when handle periodic job with github driver
Now, if we config periodic pipeline with github driver, the periodic jobs
won't be triggered and an AttributeError will be raised. That is because
the TimerTriggerEvent object don't have "commits" attribute which is used
in the zuul.driver.github.githubconnection.getPushedFileNames method.
We don't need the attribute assignment in periodic job, so this change
skipp it.

Change-Id: Ie1e85cc57ada17eacc44fb084f5f3f17aa8e9c45