Switch to stestr
stestr is maintained, while testrepository is not. OpenStack is
migrating to stestr, go ahead and follow suit.

The --no-subunit-trace flag is added to keep the quiet test run
behavior. If we remove it, stestr will by default print the name of each
test as it starts it.

As part of this, shift the default values for STDOUT capture into the
tox.ini file. This should preserve the current behavior of tox doing
capture and running tests directly not capturing.

Remove the depend on python-subunit... stestr takes care of that for us.

Change-Id: I90ae343a5db966f74e888bb98aab0bf39dcdd2cb