Register .json routes with and without suffix
In Zuul v2, we started with a /status.json endpoint, and eventually also
added a /status endpoint as well. In general, there are people who like
api endpoints both with and without suffixes.

Luckily it's pretty simple to register everything both ways.

For any path in our list of routes, if it ends with .json, simply
register it twice.

Change-Id: I4baf33fdacaf46943fbd192743551bb27bd618de