Fall back on parent jobs from default branch
If a job is defined in a project with "master" and "stable" branches,
and a user wants to inherit from that job in a project with only a
"devel" branch, Zuul will not run the job because there are no matching
variants.  In this case, look for variants on the default branch
of the parent job's project that match.  The idea is that someone defined
the child job to run on some branches, so they probably want some version
of the parent job.

Change-Id: I09867e54193fce4b4c9afcd8a1c7444e3ad2f462