WIP: Fall back on default branch for jobs
This expands the job variant fallback feature to operate on all jobs,
not just parents.  It is still the case that a project-pipeline variant
must match in order for a job to run, so a user can say "run this job
only on changes to this branch" in that manner.  However, if they have
added a job to a project centrally, and are relying on there being no
job definition which matches a given branch in order to prevent that
job from running on that branch, this will no longer work.  Zuul will
likely fall back on the job variants which match the default branch in
that case.  It may be difficult to see that the wrong version of a job
is running.

This needs the following two tests to be updated, as they rely on the
old behavior to detect no matching jobs.

tests.unit.test_v3.TestPragmaMultibranch.test_no_branch_matchers
tests.unit.test_v3.TestPragmaMultibranch.test_supplied_branch_matchers

Change-Id: I5f3e31bc8fd353c148af7b50bac6680a7bae1c88