[WIP] zuul web: add admin endpoint, enqueue commands
Create a specific, optional web admin interface (deactivated by
default).
Allow a user to trigger an `enqueue` or an `enqueue_ref` from this
web interface. Since these actions should be reserved to privileged
users, it is implied that this new interface will be run on a
restricted address or protected behind an HTTP service.

Change-Id: I6495458b8519339038ae0167e5669e29871fbcc3
TODO: add tests