zuul web: add admin endpoint, enqueue & autohold commands
Create a specific, optional web admin route under /api/admin
The route is deactivated by default, but can be enabled by setting
`enable_admin_endpoint` to True in the zuul.conf file.

Allow a user to trigger an `enqueue`, an `enqueue_ref` or an `autohold`
from this web interface. Since these actions should be reserved to
privileged users, it is implied that this route will be protected
behind an HTTP proxy.

Change-Id: I6495458b8519339038ae0167e5669e29871fbcc3