Update patch set 2
Patch Set 2: Code-Review+2

This *is* being executed by the tests, but since the tests don't have installations for projects, the installation map is empty, and the .get() returns None, which is okay, because without installations, you would need to use the anonymous client anyway, and that's what happens.  So the error only occurs when there are actual installations involved.

We should look into improving our tests for that, but I don't have a good idea for that right now.

Patch-set: 2
Label: Code-Review=+2