Update patch set 1
Patch Set 1: Code-Review-1

I think most of this approach may work -- accumulating errors as we load, and checking at the end.  However, they are currently being accumulated on the tenant, and during dynamic reconfiguration, the "production" tenant is used.  Only a new layout is created.  That means that if we encountered an error in a dynamic reload, it will persist on the tenant.  It looks like they would simply accumulate.

Perhaps we should accumulate them on the layout object instead.

Also, I realize that the configloader file is complicated right now.  Many of the static methods should just be converted into real instance methods on normal classes.  The entrypoint to the whole process should probably be a dynamically created parser object as well.  I haven't had as much time to work on converting those classes as I would like (some have been -- the JobParser for example).  But perhaps when that's done, it would make sense to accumulate the errors on the dynamic parser object instead?

We should also probably add a test case which verifies that errors from a dynamic layout don't show up on the tenant load errors.

There are also some cases where we perform early-binding of objects that this approach could be problematic.  For example, if there is an error that causes the default nodeset not to be loaded, that could cascade to cause errors in every job.  It's probably worth considering the performance impact of that at openstack scale.

Patch-set: 1
Label: Code-Review=-1