Update patch set 8
Patch Set 8: Code-Review-1

(1 comment)

-1 for the comment in the commit message.

Structurally, I'd like it if all of the github specific code could live in the driver itself.  So when zuul-web starts and configures its connections, it can interrogate all of the connections and find out if they have code they need to load, and if so, establish a route for them which calls into the driver.

Hopefully the requirements for that are separate enough that we can do that without entangling too much of the driver with the asyncio stuff.  Maybe it can be implemented as a new feature of drivers (like Source, Trigger, Reporter -- WebHook).  So the connection init for web would iterate over all the connections, and if a connection's driver has a WebHook, instantiate it and attach it to a route.

I think we can do that as a follow-up.  The most important thing for this is to get the public API right.

Patch-set: 8