Allow .yml extensions for user config files
For some people the '.yml' extension seems to make them happier than the
'.yaml' extension. Ansible allows both extensions for yaml files, so
ansible users coming to zuul might be confused that zuul only accepts
.zuul.yaml causing errors or confusion such as are addressed by:

  https://review.openstack.org/#/c/540669

Update config loading to look for .yml as well. Update the docs just
with a note saying that .yml is accepted, but leave the rest of the docs
just mentioning .yaml so that the docs don't read poorly.

Change-Id: I5039815880536ed39392980bab444ab1614c2733