Update patch set 2
Patch Set 2:

> I think this is an interesting approach. However directly invoking
 > a third party function within the zuul process sounds a bit awkward
 > to me. This has a few implications I consider problematic.
 
For a first proposal I have used the same approach than the one
used (in zuul2) to pass custom parameters to jobs
"includes: [{"python-file": "zuul_functions.py"}]". The same sort
of loading of python code was used.

 > We would get the data back how it’s in our internal data structure.
 > I don’t think that should be part of our api. This could make later
 > improvements harder.
 
The data returned by the external function must be, indeed, a list of
correct project items. The format is the same
as the tenant file. The returned data format must evolve when the config
tenant api evolve. I don't understand what do you mean by "This could
make later improvements harder."

 > Further also internal data structures/functions would be made
 > available to the script which also makes updating harder.
 
I've checked again and, as far as see, from the script, access
to zuul data/functions is not possible and even it should be possible
that should not be recommended manipulate or call them.

 > So if we want dynamic loading of tenant configuration possible I
 > think we should at least separate this out into an external process
 > with a clear api (e.g. just the tenant config spec which is an
 > official api) and maybe read yaml-generated tenant config via
 > stdout. I’m not sure if we should support loading project lists but
 > maybe just dynamically generated toplevel tenant config trees.

Yes, this approach can be changed to call an external command and let
the configloader read the command stdout (a valid yaml tenants
description file). What about having as an alternative to "tenant-config"
a new "tenant-config-script" that the configloader can call ? Would
you like I propose a new PS with this new approach ?

Patch-set: 2
Label: Verified=0