Allow deleting autoholds
It happens at times where users add autoholds in anticipation of a
failure, only to find that the job works and they will no longer need
held nodes.

Allow deleting them by passing the same args. We should possibly
consider exposing the autohold ID and allowing delete by that /
returning it on create, but for now, this should work fine.

Change-Id: I55cf835aae06b7366f7bedef5116e38398a408bb